Implemented main gameplay feature
On desktop, the spacebar will make the player jump as long as the back wheel of the bike has contact with the ground.

If there is no contact, the bike will tilt backwards. It is possible to configure the tilt torque and max speed (altough it is a bit buggy).

If the bike's front wheel has no contact with the ground, a slight tilt forwards is automatically applide. This allows better control for the player.