Add ground elements to "Ground" layer
The code to detect wether or not the player is in contact with the ground relies on knowing if the player is colliding with anything in the Ground layer. This is used to know when the player is allowed to jump (only when the back wheel is touching something classified as "Ground", on the ground layer).