Better "grounded" detection
Now in order to detect contact with the ground there will be a ray cast to decide what collider to use as "ground". After that, we get the minimum distance to that collider and use a threashold value. This should be more roboust than detecting direct contact.

Now collision detection is continuous for both wheels as well.